rename: gameManager interface를 voteManager로 바꿔 역할에 맞게 네이밍을 확실하게 진행
