+++ BE/src/game/usecase/allocate-user-role/allocate.user-role.service.ts
@@ -7,7 +7,7 @@ import { MAFIA_ROLE } from '../../mafia-role';
 import { Transactional } from 'typeorm-transactional';
 import { JOB_FACTORY, JobFactory } from './job.factory';
 import { GAME_HISTORY_REPOSITORY, GameHistoryRepository } from '../../repository/game-history.repository';
-import { GAME_MANAGER, GameManager } from '../game-manager/game-manager';
+import { GAME_MANAGER, VoteManager } from '../game-manager/vote-manager';
 
 @Injectable()
 export class AllocateUserRoleService implements AllocateUserRoleUsecase {
@@ -18,7 +18,7 @@ export class AllocateUserRoleService implements AllocateUserRoleUsecase {
     @Inject(GAME_HISTORY_REPOSITORY)
     private readonly gameHistoryRepository: GameHistoryRepository<GameHistoryEntity, number>,
     @Inject(GAME_MANAGER)
-    private readonly gameManager: GameManager,
+    private readonly gameManager: VoteManager,
   ) {
   }
