+++ BE/src/game/usecase/game-manager/vote.mafia.service.ts
@@ -1,15 +1,15 @@
 import { VoteMafiaUsecase } from './vote.mafia.usecase';
 import { Inject, Injectable } from '@nestjs/common';
 import { GameRoom } from '../../../game-room/entity/game-room.model';
-import { GAME_MANAGER, GameManager } from './game-manager';
+import { VOTE_MANAGER, VoteManager } from './vote-manager';
 import { VOTE_STATE } from '../../vote-state';
 
 @Injectable()
 export class VoteMafiaService implements VoteMafiaUsecase {
 
   constructor(
-    @Inject(GAME_MANAGER)
-    private readonly gameManager: GameManager,
+    @Inject(VOTE_MANAGER)
+    private readonly gameManager: VoteManager,
   ) {
   }
